% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_protein_report}
\alias{generate_protein_report}
\title{Generate Plasmix Proteomics report}
\usage{
generate_protein_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated result to generate report
}
\examples{
# 加载示例数据
# 运行函数

}
