% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_metrics.R
\name{get_sample_colors}
\alias{get_sample_colors}
\title{Get color mapping for samples}
\usage{
get_sample_colors(samples)
}
\arguments{
\item{samples}{Vector of sample names}
}
\value{
Named vector of colors
}
\description{
Get color mapping for samples
}
