% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pipelines.R
\name{qc_allmetrics}
\alias{qc_allmetrics}
\title{Calculating all QC metrics}
\usage{
qc_allmetrics(expr_dt, meta_dt, output_dir = NULL, plot = TRUE)
}
\arguments{
\item{expr_dt}{A generic expression table (Feature level)}

\item{meta_dt}{A metadata file}

\item{output_dir}{A directory for results}

\item{plot}{if True, a plot will be output.}
}
\description{
Calculating all QC metrics
}
