% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_protein_report}
\alias{generate_protein_report}
\title{Generate Protein Quality Report (Chinese Version)}
\usage{
generate_protein_report(
  qc_result,
  report_template = NULL,
  report_dir = NULL,
  report_name = NULL,
  batch_name = NULL
)
}
\arguments{
\item{qc_result}{list, the output from qc_conclusion()}

\item{report_template}{character, path to a .docx file to use as style template (optional)}

\item{report_dir}{character, directory to save the report}

\item{report_name}{character, filename of the report}

\item{batch_name}{character, optional, manual batch name. If NULL, tries to infer from metadata.}
}
\description{
Generates a report following the ShenKang Plasmix template structure.
}
